+++ src/styles/tokens.ts
@@ -2,6 +2,11 @@
 export const cssTokenNames = [
   '--foreground',
   '--background',
+  '--primary',
+  '--secondary',
+  '--confirm',
+  '--danger',
+  '--overlay',
 ] as const;
 
 export type CssTokenName = typeof cssTokenNames[number];